Fix ANR due to blocking call on main thread (fix #713)
