+++ News-Android-App/src/main/AndroidManifest.xml
@@ -122,9 +122,8 @@
                 android:name="android.accounts.AccountAuthenticator"
                 android:resource="@xml/authenticator" />
         </service>
-        <service android:name=".services.OwnCloudSyncService" />
         <service
-            android:name=".services.OwnCloudSettingsSyncService"
+            android:name=".services.OwnCloudSyncService"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.content.SyncAdapter" />
